backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
gunicorn>=21.2
cloudinary>=1.36
django-cloudinary-storage>=0.3.0
dj-database-url>=2.1
psycopg2-binary>=2.9
